Make module selection a boolean dict
This allows page.modules to fall back to the default, instead of
overriding it. This saves you from having to specify all wanted modules
(including the defaults) if you want anything beyond the defaults. Also,
this mean that default modules can be added without needing to change
all module declarations.